Convert key to int instead of tuple and store score totals and seen sequences in arrays instead of map/set

+++ AoC2024/Day22/Day22.h
@@ -21,6 +21,6 @@ namespace AoC2024 {
         }
 
     private:
-        std::vector<std::int64_t> input = {};
+        std::vector<std::int32_t> input = {};
     };
 }
